Moved nouns auction functions to their own file.
